chore: refactor anon Fx to spacebarKey Fx. Closes #15
